+++ src/views/Stake/Stake.tsx
@@ -1,39 +1,21 @@
 import "./Stake.scss";
 
-import { t } from "@lingui/macro";
-import { Grid, Zoom } from "@material-ui/core";
-import { MetricCollection, Paper } from "@olympusdao/component-library";
-import { memo, useState } from "react";
+import { memo } from "react";
+import { useHistory } from "react-router-dom";
+import { useWeb3Context } from "src/hooks";
+import { usePathForNetwork } from "src/hooks/usePathForNetwork";
 
-import { CurrentIndex, StakingAPY, TotalValueDeposited } from "../TreasuryDashboard/components/Metric/Metric";
 import ExternalStakePools from "./components/ExternalStakePools/ExternalStakePools";
-import RebaseTimer from "./components/RebaseTimer/RebaseTimer";
-import { StakeActionArea } from "./components/StakeActionArea/StakeActionArea";
+import { StakeArea } from "./components/StakeArea/StakeArea";
 
 const Stake: React.FC = () => {
-  const [isZoombed, setIsZoomed] = useState(false);
+  const history = useHistory();
+  const { networkId } = useWeb3Context();
+  usePathForNetwork({ pathName: "stake", networkID: networkId, history });
 
   return (
     <div id="stake-view">
-      <Zoom in onEntered={() => setIsZoomed(true)}>
-        <Paper headerText={t`Single Stake (3, 3)`} subHeader={<RebaseTimer />}>
-          <Grid container direction="column" spacing={2}>
-            <Grid item>
-              <MetricCollection>
-                <StakingAPY className="stake-apy" />
-
-                <TotalValueDeposited className="stake-tvl" />
-
-                <CurrentIndex className="stake-index" />
-              </MetricCollection>
-            </Grid>
-
-            <div className="staking-area">
-              <StakeActionArea isZoomed={isZoombed} />
-            </div>
-          </Grid>
-        </Paper>
-      </Zoom>
+      <StakeArea />
 
       {/* NOTE (appleseed-olyzaps) olyzaps disabled until v2 contracts */}
       {/* <ZapCta /> */}
