+++ src/views/Stake/components/StakeArea/components/RebaseTimer/RebaseTimer.scss
