+++ src/views/Stake/components/StakeArea/components/RebaseTimer/RebaseTimer.tsx
@@ -5,7 +5,7 @@ import { Box, Typography } from "@material-ui/core";
 import { Skeleton } from "@material-ui/lab";
 import React from "react";
 import { prettifySeconds } from "src/helpers";
-import { useNextRebaseDate } from "src/views/Stake/hooks/useNextRebaseDate";
+import { useNextRebaseDate } from "src/views/Stake/components/StakeArea/components/RebaseTimer/hooks/useNextRebaseDate";
 
 const RebaseTimer: React.FC = () => {
   const { data: nextRebaseDate } = useNextRebaseDate();
