+++ src/views/Stake/components/StakeArea/components/RebaseTimer/hooks/useNextRebaseDate.ts
@@ -5,13 +5,13 @@ import { parseBigNumber } from "src/helpers";
 import { useStakingContract } from "src/hooks/useContract";
 import { useStaticProvider } from "src/hooks/useStaticProvider";
 
-export const nextRebaseDate = () => ["useNextRebaseDate"];
+export const nextRebaseDateQueryKey = () => ["useNextRebaseDate"];
 
 export const useNextRebaseDate = () => {
   const provider = useStaticProvider(NetworkId.MAINNET);
   const contract = useStakingContract(STAKING_ADDRESSES[NetworkId.MAINNET], provider);
 
-  return useQuery<Date, Error>(nextRebaseDate(), async () => {
+  return useQuery<Date, Error>(nextRebaseDateQueryKey(), async () => {
     const secondsToRebase = await contract.secondsToNextEpoch();
 
     const parsedSeconds = parseBigNumber(secondsToRebase, 0);
