+++ src/views/Stake/components/StakeArea/components/StakeActionArea/components/StakeNextRebaseAmount.tsx
